app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
asyncpg==0.29.0
redis==5.0.3
pydantic==2.6.4
psycopg2-binary==2.9.9